+++ packages/profile-sync-controller/src/sdk/authentication-jwt-bearer/flow-srp.ts
@@ -15,6 +15,7 @@ import {
   authorizeOIDC,
   getCustomerServiceToken,
   getNonce,
+  getOidcToken,
   getUserProfileLineage,
   pairProfiles,
   pairSocialIdentifier,
@@ -28,6 +29,8 @@ import type {
   IBaseAuth,
   LoginIdentifierType,
   LoginResponse,
+  OidcTokenAudience,
+  OidcTokenClaim,
   PairSocialIdentifierParams,
   SrpLoginTag,
   UserProfile,
@@ -193,6 +196,15 @@ export class SRPJwtBearerAuth implements IBaseAuth {
   async getCustomerServiceToken(entropySourceId?: string): Promise<string> {
     const accessToken = await this.getAccessToken(entropySourceId);
     return await getCustomerServiceToken(this.#config.env, accessToken);
+  }
+
+  async getOidcToken(
+    claims: OidcTokenClaim[],
+    audience: OidcTokenAudience,
+    entropySourceId?: string,
+  ): Promise<string> {
+    const accessToken = await this.getAccessToken(entropySourceId);
+    return await getOidcToken(this.#config.env, accessToken, claims, audience);
   }
 
   async pairSocialIdentifier(